Add ChatBubble dynamic background
