+++ GameJamProject/Assets/ChatBubble.png.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: 0a96cec25dc8a4f47b08d9c190ddc8b6
+guid: c12a24caba0eb4faa88e1e624e0cd60c
 TextureImporter:
   internalIDToNameTable:
   - first:
-      213: 7136138100767428815
-    second: bad_money_button_0
+      213: 8571661363573756414
+    second: ChatBubble_0
   externalObjects: {}
   serializedVersion: 13
   mipmaps:
@@ -113,13 +113,13 @@ TextureImporter:
     serializedVersion: 2
     sprites:
     - serializedVersion: 2
-      name: bad_money_button_0
+      name: ChatBubble_0
       rect:
         serializedVersion: 2
-        x: 0
-        y: 0
-        width: 64
-        height: 32
+        x: 23
+        y: 27
+        width: 98
+        height: 62
       alignment: 0
       pivot: {x: 0, y: 0}
       border: {x: 0, y: 0, z: 0, w: 0}
@@ -128,8 +128,8 @@ TextureImporter:
       physicsShape: []
       tessellationDetail: -1
       bones: []
-      spriteID: fc09f543bc7a80360800000000000000
-      internalID: 7136138100767428815
+      spriteID: ef111c67898a4f670800000000000000
+      internalID: 8571661363573756414
       vertices: []
       indices: 
       edges: []
@@ -148,7 +148,7 @@ TextureImporter:
     spriteCustomMetadata:
       entries: []
     nameFileIdTable:
-      bad_money_button_0: 7136138100767428815
+      ChatBubble_0: 8571661363573756414
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
