+++ GameJamProject/Assets/DialogueSpawnScript.cs
@@ -8,7 +8,7 @@ public class DialogueSpawnScript : MonoBehaviour
     void Start()
     {
         spawnRate = 5;
-        timer = 3;
+        timer = 4;
     }
 
     // Update is called once per frame
@@ -20,7 +20,6 @@ public class DialogueSpawnScript : MonoBehaviour
         else {
             GameObject clone = Instantiate(bubble, new Vector3(transform.position[0] - 3.5f, transform.position[1] + 0.5f, 0), transform.rotation);
             clone.transform.parent = transform;
-            //setText(clone);
             timer = 0;
         }
     }
